pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "segmentation-suite"
version = "1.0.0"
description = "MOSS - Microscopy Oriented Segmentation with Supervision. Interactive segmentation tool with U-Net training and prediction."
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "PyQt6>=6.4.0",
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.7.0",
    "scikit-image>=0.20.0",
    "tifffile>=2023.1.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.65.0",
    "scipy>=1.10.0",
    "Pillow>=9.0.0",
    "websockets>=12.0",
]

[project.scripts]
segmentation-suite = "segmentation_suite.launcher:main"

[project.urls]
Homepage = "https://github.com/yourusername/segmentation-suite"

[tool.setuptools.packages.find]
where = ["."]
include = ["segmentation_suite*"]
